feat: membership fixes for org
